fix(tests): green CI on CPU + faster default suite
Real bug fixes
--------------
- src/mime/control/kinematics/ik.py: cast the fori_loop carry back to
  q_init.dtype each iteration. With another test having flipped
  jax_enable_x64 on earlier in the process, pose_to_matrix /
  jnp.linalg.solve silently upcast the float32 carry to float64 and
  fori_loop strict-type checking failed on the second iteration.
- experiments/ar4_helical_drive/control/controller.py: same fix on
  the controller's IK fori_loop.
- src/mime/nodes/environment/lbm/d3q19.py: pin init_equilibrium to
  float32 (jnp.ones/zeros without dtype were picking up the global
  default and producing float64 when x64 leaked in).
- tests/nodes/test_magnetic_response.py: the rotation-changes-torque
  assertion used a geometry where m × B = 0 in both identity and
  rotated cases (B along x, rotation about z). Tilt about y so the
  body's e1 leaves the B direction; relax atol to 1e-9 (s2 is small
  but many orders above s1).

CI hygiene
----------
- pyproject.toml: register the `gpu` marker so the FVM verification
  warnings stop. tests/conftest.py auto-skips gpu-marked tests when
  no CUDA device is visible (the CPU path NaNs on those tests).
- Mark the following tests `slow` so the default `not slow` CI filter
  skips them: large LBM convergence sweeps, Ladd-cylinder Bouzidi
  torque, MLP-clamp data-loading tests, b0 chain stability, the
  PermanentMagnet far-field + misalignment benchmarks, the new-chain
  graph smoke test, and the AR4 orientation/safety tests that each
  pay a 25-40 s JIT cost. CI suite drops from ~14 min to ~9 min on
  CPU; full coverage available with `pytest -m 'slow or not slow'`.

Test-bug fixes
--------------
- AR4 orientation/position/safety tests override CONTROL_TARGET_ALPHA
  to 0.2 in their params dict. params.py uses 0.005 for the live
  experiment (slow LP for jitter rejection), but those tests run
  only 200 iterations and don't converge at that rate — the
  fori_loop dtype bug masked this until now.

+++ src/mime/nodes/environment/lbm/d3q19.py
@@ -404,8 +404,11 @@ def init_equilibrium(
     -------
     f : (nx, ny, nz, 19) float32
     """
-    rho = jnp.ones((nx, ny, nz)) * density
-    u = jnp.zeros((nx, ny, nz, 3))
+    # Explicit float32 — without dtype, jnp.ones/zeros pick up JAX's
+    # default dtype which becomes float64 if any other test in the
+    # process has flipped jax_enable_x64 on.
+    rho = jnp.ones((nx, ny, nz), dtype=jnp.float32) * density
+    u = jnp.zeros((nx, ny, nz, 3), dtype=jnp.float32)
     u = u.at[..., 0].set(velocity[0])
     u = u.at[..., 1].set(velocity[1])
     u = u.at[..., 2].set(velocity[2])
